[SCRUM-59] refactor: 위경도 조회 시그니처 BigDecimal로 변경

+++ src/main/java/com/sprint/otboo/weather/repository/WeatherLocationRepository.java
@@ -1,12 +1,13 @@
 package com.sprint.otboo.weather.repository;
 
 import com.sprint.otboo.weather.entity.WeatherLocation;
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WeatherLocationRepository extends JpaRepository<WeatherLocation, UUID> {
 
-    Optional<WeatherLocation> findFirstByLatitudeAndLongitude(double latitude, double longitude);
+    Optional<WeatherLocation> findFirstByLatitudeAndLongitude(BigDecimal latitude, BigDecimal longitude);
     Optional<WeatherLocation> findFirstByXAndY(int x, int y);
 }
